+++ pkg/cloud/services/route53/route53.go
@@ -3,6 +3,7 @@ package route53
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -31,18 +32,7 @@ func (s *Service) DeleteRoute53(ctx context.Context) error {
 		return microerror.Mask(err)
 	}
 
-	// We need to delete all records first before we can delete the hosted zone
-	err = s.changeClusterIngressRecords(ctx, actionDelete)
-	if IsNotFound(err) {
-		// Entry does not exist, fall through
-	} else if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = s.changeClusterAPIRecords(ctx, actionDelete)
-	if IsNotFound(err) {
-		// Entry does not exist, fall through
-	} else if err != nil {
+	if err := s.deleteClusterRecords(ctx, hostedZoneID); err != nil {
 		return microerror.Mask(err)
 	}
 
@@ -67,9 +57,9 @@ func (s *Service) ReconcileRoute53(ctx context.Context) error {
 	s.scope.V(2).Info("Reconciling hosted DNS zone")
 
 	// Describe or create.
-	_, err := s.describeClusterHostedZone(ctx)
+	hostedZoneID, err := s.describeClusterHostedZone(ctx)
 	if IsHostedZoneNotFound(err) {
-		err = s.createClusterHostedZone(ctx)
+		hostedZoneID, err = s.createClusterHostedZone(ctx)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -82,11 +72,11 @@ func (s *Service) ReconcileRoute53(ctx context.Context) error {
 		return microerror.Mask(err)
 	}
 
-	if err := s.changeClusterAPIRecords(ctx, actionUpsert); err != nil {
+	if err := s.changeClusterAPIRecords(ctx, hostedZoneID, actionUpsert); err != nil {
 		return microerror.Mask(err)
 	}
 
-	if err := s.changeClusterIngressRecords(ctx, actionUpsert); err != nil {
+	if err := s.changeClusterIngressRecords(ctx, hostedZoneID, actionUpsert); err != nil {
 		return microerror.Mask(err)
 	}
 
@@ -96,7 +86,7 @@ func (s *Service) ReconcileRoute53(ctx context.Context) error {
 func (s *Service) describeClusterHostedZone(ctx context.Context) (string, error) {
 	// Search host zone by DNSName
 	input := &route53.ListHostedZonesByNameInput{
-		DNSName: aws.String(fmt.Sprintf("%s.%s", s.scope.Name(), s.scope.BaseDomain())),
+		DNSName: aws.String(s.scope.ClusterDomain()),
 	}
 	out, err := s.Route53Client.ListHostedZonesByNameWithContext(ctx, input)
 	if err != nil {
@@ -107,22 +97,59 @@ func (s *Service) describeClusterHostedZone(ctx context.Context) (string, error)
 		return "", microerror.Mask(hostedZoneNotFoundError)
 	}
 
-	if *out.HostedZones[0].Name != fmt.Sprintf("%s.%s.", s.scope.Name(), s.scope.BaseDomain()) {
+	if *out.HostedZones[0].Name != fmt.Sprintf("%s.", s.scope.ClusterDomain()) {
 		return "", microerror.Mask(hostedZoneNotFoundError)
 	}
 
 	return *out.HostedZones[0].Id, nil
 }
 
-func (s *Service) listClusterNSRecords(ctx context.Context) ([]*route53.ResourceRecord, error) {
-	hostZoneID, err := s.describeClusterHostedZone(ctx)
-	if err != nil {
-		return nil, microerror.Mask(err)
+func (s *Service) deleteClusterRecords(ctx context.Context, hostedZoneID string) error {
+	input := &route53.ListResourceRecordSetsInput{
+		HostedZoneId: aws.String(hostedZoneID),
 	}
 
+	output, err := s.Route53Client.ListResourceRecordSetsWithContext(ctx, input)
+	if err != nil {
+		return wrapRoute53Error(err)
+	}
+
+	recordsToDelete := &route53.ChangeResourceRecordSetsInput{
+		HostedZoneId: aws.String(hostedZoneID),
+		ChangeBatch: &route53.ChangeBatch{
+			Changes: []*route53.Change{},
+		},
+	}
+
+	for _, recordSet := range output.ResourceRecordSets {
+		if strings.TrimSuffix(*recordSet.Name, ".") == s.scope.ClusterDomain() {
+			// We cannot delete those entries, they get automatically cleaned up when deleting the hosted zone
+			continue
+		}
+
+		recordsToDelete.ChangeBatch.Changes = append(recordsToDelete.ChangeBatch.Changes, &route53.Change{
+			Action:            aws.String(actionDelete),
+			ResourceRecordSet: recordSet,
+		})
+	}
+
+	if len(recordsToDelete.ChangeBatch.Changes) == 0 {
+		// Nothing to delete
+		return nil
+	}
+
+	_, err = s.Route53Client.ChangeResourceRecordSetsWithContext(ctx, recordsToDelete)
+	if err != nil {
+		return wrapRoute53Error(err)
+	}
+
+	return nil
+}
+
+func (s *Service) listClusterNSRecords(ctx context.Context, hostedZoneID string) ([]*route53.ResourceRecord, error) {
 	// First entry is always NS record
 	input := &route53.ListResourceRecordSetsInput{
-		HostedZoneId: aws.String(hostZoneID),
+		HostedZoneId: aws.String(hostedZoneID),
 		MaxItems:     aws.String("1"),
 	}
 
@@ -133,26 +160,21 @@ func (s *Service) listClusterNSRecords(ctx context.Context) ([]*route53.Resource
 	return output.ResourceRecordSets[0].ResourceRecords, nil
 }
 
-func (s *Service) changeClusterAPIRecords(ctx context.Context, action string) error {
+func (s *Service) changeClusterAPIRecords(ctx context.Context, hostedZoneID, action string) error {
 	s.scope.Info(s.scope.APIEndpoint())
 	if s.scope.APIEndpoint() == "" {
 		s.scope.Info("API endpoint is not ready yet.")
 		return aws.ErrMissingEndpoint
 	}
 
-	hostZoneID, err := s.describeClusterHostedZone(ctx)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
 	input := &route53.ChangeResourceRecordSetsInput{
-		HostedZoneId: aws.String(hostZoneID),
+		HostedZoneId: aws.String(hostedZoneID),
 		ChangeBatch: &route53.ChangeBatch{
 			Changes: []*route53.Change{
 				{
 					Action: aws.String(action),
 					ResourceRecordSet: &route53.ResourceRecordSet{
-						Name: aws.String(fmt.Sprintf("api.%s.%s", s.scope.Name(), s.scope.BaseDomain())),
+						Name: aws.String(fmt.Sprintf("api.%s", s.scope.ClusterDomain())),
 						Type: aws.String("A"),
 						TTL:  aws.Int64(300),
 						ResourceRecords: []*route53.ResourceRecord{
@@ -166,37 +188,28 @@ func (s *Service) changeClusterAPIRecords(ctx context.Context, action string) er
 		},
 	}
 
-	_, err = s.Route53Client.ChangeResourceRecordSetsWithContext(ctx, input)
-	if err != nil {
+	if _, err := s.Route53Client.ChangeResourceRecordSetsWithContext(ctx, input); err != nil {
 		return wrapRoute53Error(err)
 	}
 	return nil
 }
 
-func (s *Service) changeClusterIngressRecords(ctx context.Context, action string) error {
-	hostZoneID, err := s.describeClusterHostedZone(ctx)
+func (s *Service) changeClusterIngressRecords(ctx context.Context, hostedZoneID, action string) error {
+	ingressIP, err := s.getIngressIP(ctx)
 	if err != nil {
 		return microerror.Mask(err)
-	}
-
-	var ingressIP string
-	if action == actionUpsert { // Avoid looking up IP via k8s client when deleting
-		ingressIP, err = s.getIngressIP(ctx)
-		if err != nil {
-			return microerror.Mask(err)
-		} else if ingressIP == "" {
-			return nil
-		}
+	} else if ingressIP == "" {
+		return nil
 	}
 
 	input := &route53.ChangeResourceRecordSetsInput{
-		HostedZoneId: aws.String(hostZoneID),
+		HostedZoneId: aws.String(hostedZoneID),
 		ChangeBatch: &route53.ChangeBatch{
 			Changes: []*route53.Change{
 				{
 					Action: aws.String(action),
 					ResourceRecordSet: &route53.ResourceRecordSet{
-						Name: aws.String(fmt.Sprintf("ingress.%s.%s", s.scope.Name(), s.scope.BaseDomain())),
+						Name: aws.String(fmt.Sprintf("ingress.%s", s.scope.ClusterDomain())),
 						Type: aws.String("A"),
 						TTL:  aws.Int64(TTL),
 						ResourceRecords: []*route53.ResourceRecord{
@@ -209,12 +222,12 @@ func (s *Service) changeClusterIngressRecords(ctx context.Context, action string
 				{
 					Action: aws.String(action),
 					ResourceRecordSet: &route53.ResourceRecordSet{
-						Name: aws.String(fmt.Sprintf("*.%s.%s", s.scope.Name(), s.scope.BaseDomain())),
+						Name: aws.String(fmt.Sprintf("*.%s", s.scope.ClusterDomain())),
 						Type: aws.String("CNAME"),
 						TTL:  aws.Int64(300),
 						ResourceRecords: []*route53.ResourceRecord{
 							{
-								Value: aws.String(fmt.Sprintf("ingress.%s.%s", s.scope.Name(), s.scope.BaseDomain())),
+								Value: aws.String(fmt.Sprintf("ingress.%s", s.scope.ClusterDomain())),
 							},
 						},
 					},
@@ -223,8 +236,7 @@ func (s *Service) changeClusterIngressRecords(ctx context.Context, action string
 		},
 	}
 
-	_, err = s.Route53Client.ChangeResourceRecordSetsWithContext(ctx, input)
-	if err != nil {
+	if _, err := s.Route53Client.ChangeResourceRecordSetsWithContext(ctx, input); err != nil {
 		return wrapRoute53Error(err)
 	}
 	return nil
@@ -251,24 +263,24 @@ func (s *Service) describeBaseHostedZone(ctx context.Context) (string, error) {
 }
 
 func (s *Service) changeClusterNSDelegation(ctx context.Context, action string) error {
-	hostZoneID, err := s.describeBaseHostedZone(ctx)
+	hostedZoneID, err := s.describeBaseHostedZone(ctx)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
-	records, err := s.listClusterNSRecords(ctx)
+	records, err := s.listClusterNSRecords(ctx, hostedZoneID)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
 	input := &route53.ChangeResourceRecordSetsInput{
-		HostedZoneId: aws.String(hostZoneID),
+		HostedZoneId: aws.String(hostedZoneID),
 		ChangeBatch: &route53.ChangeBatch{
 			Changes: []*route53.Change{
 				{
 					Action: aws.String(action),
 					ResourceRecordSet: &route53.ResourceRecordSet{
-						Name:            aws.String(fmt.Sprintf("%s.%s", s.scope.Name(), s.scope.BaseDomain())),
+						Name:            aws.String(s.scope.ClusterDomain()),
 						Type:            aws.String("NS"),
 						TTL:             aws.Int64(TTL),
 						ResourceRecords: records,
@@ -285,11 +297,11 @@ func (s *Service) changeClusterNSDelegation(ctx context.Context, action string) 
 	return nil
 }
 
-func (s *Service) createClusterHostedZone(ctx context.Context) error {
+func (s *Service) createClusterHostedZone(ctx context.Context) (string, error) {
 	now := time.Now()
 	input := &route53.CreateHostedZoneInput{
 		CallerReference: aws.String(now.UTC().String()),
-		Name:            aws.String(fmt.Sprintf("%s.%s.", s.scope.Name(), s.scope.BaseDomain())),
+		Name:            aws.String(fmt.Sprintf("%s.", s.scope.ClusterDomain())),
 	}
 
 	if s.scope.ManagementCluster() != "" {
@@ -297,11 +309,12 @@ func (s *Service) createClusterHostedZone(ctx context.Context) error {
 			Comment: aws.String(fmt.Sprintf("management_cluster: %s", s.scope.ManagementCluster())),
 		}
 	}
-	_, err := s.Route53Client.CreateHostedZoneWithContext(ctx, input)
+	output, err := s.Route53Client.CreateHostedZoneWithContext(ctx, input)
 	if err != nil {
-		return wrapRoute53Error(err)
+		return "", wrapRoute53Error(err)
 	}
-	return nil
+
+	return *output.HostedZone.Id, nil
 }
 
 func (s *Service) deleteClusterHostedZone(ctx context.Context, hostedZoneID string) error {